chore(mobile): dedup withAlpha, formatTrackDuration, ICON_STROKE
Extract three duplicated utilities into shared modules:
- src/lib/color.ts (withAlpha)
- src/lib/group-tracks-by-album.ts (formatTrackDuration alongside formatAlbumDuration)
- src/theme/icons.ts (ICON_STROKE)
