+++ apps/mobile/src/lib/color.ts
@@ -1,0 +1,7 @@
+export function withAlpha(hex: string, alpha: number): string {
+  const m = /^#([0-9a-f]{6})$/i.exec(hex);
+  if (!m) return hex;
+  const a = Math.max(0, Math.min(1, alpha));
+  const aHex = Math.round(a * 255).toString(16).padStart(2, '0');
+  return `#${m[1]}${aHex}`;
+}
