+++ apps/mobile/src/theme/icons.ts
@@ -1,0 +1,1 @@
+export const ICON_STROKE = 2.5;
